Remove `.double` from `perplexity_compute`
MPS requires F32 but `.double` casts the output back to F64 which
causes it to fail

+++ torcheval/metrics/functional/text/perplexity.py
@@ -114,7 +114,7 @@ def _perplexity_compute(
     num_total: torch.Tensor,
 ) -> torch.Tensor:
 
-    return torch.exp(sum_log_probs / num_total).double()
+    return torch.exp(sum_log_probs / num_total)
 
 
 def _perplexity_input_check(
